Rework full specification and standalone workload specifications
This helps avoid missing references in standalone workload specifications (that refer to other chapters of the full specification) and it helps avoid repetitions in the full specification (where delete operations are only referred to from both workloads and are not replicated).

We use a toggle variable (provided by etoolbox) 'StandaloneWorkloadSpecification' to differentiate between the two cases.